Restructured and re-organised code
