+++ .github/workflows/stm32build.yml
@@ -150,6 +150,6 @@ jobs:
           MODULE_NAME : ${{ matrix.package }}
       - name: Manage PCB Versioning
         run: |
-          python .github/workflows/managePCBVersions.py -c $(echo "${{ steps.pcbversion.outputs.version }}" | sed 's/[a-zA-Z]//g') -b $(echo "${{ steps.breakingpcbversion.outputs.breaking }}" | sed 's/[a-zA-Z]//g') -fw $(echo "${{ steps.release-on-push.outputs.tag_name }}" | cut -d'-' -f2) -m ${{ matrix.package }}
+          python scripts/managePCBVersions.py -c $(echo "${{ steps.pcbversion.outputs.version }}" | sed 's/[a-zA-Z]//g') -b $(echo "${{ steps.breakingpcbversion.outputs.breaking }}" | sed 's/[a-zA-Z]//g') -fw $(echo "${{ steps.release-on-push.outputs.tag_name }}" | cut -d'-' -f2) -m ${{ matrix.package }}
         env : 
           AZURE_BLOB_QUERYSTRING : ${{ secrets.AZURE_BLOB_QUERYSTRING }}
