+++ .gitignore
@@ -7,3 +7,6 @@ githash.h
 /.NET/FFTviewer/.vs/FFTviewer/v17/.suo
 /.NET/FFTviewer/.vs/FFTviewer
 /.NET/FFTviewer/FFTviewer/obj
+
+# Compiled python files
+*.pyc
