+++ scripts/simpleConsole.py
@@ -1,0 +1,4 @@
+import subprocess
+
+def console(args):
+    return subprocess.run(args, check=True, text=True, shell=True)
